[TASK] Altered original logic for legendary items.
Legendary items don’t need to get sold. So to reflect this better in the code, we set their sellIn to Infinity which will provide future safety if any other code that alters the sellIn date should be added in the future.

As we overwrite the sellIn for legendary items, we can ditch the decreaseSellIn function and move it back to the core function and apply it to all items.